+++ R/data.R
@@ -1,0 +1,16 @@
+#' Platform and Region data from speedrun.com
+#'
+#' These datasets are saved for convenience, as they are relatively small and
+#' not expected to change frequently.
+#'
+#' @rdname platformsregions
+#' @format
+#' \describe{
+#'   \item{platforms}{A `tibble` with 99 rows and 3 variables: `id, name, released`}
+#'   \item{regions}{A `tibble` with 6 rows and 2 variables: `id, name`}
+#' }
+#'
+"platforms"
+
+#' @rdname platformsregions
+"regions"