+++ data-raw/update_data.R
@@ -1,0 +1,7 @@
+library(speedrunr)
+library(usethis)
+
+platforms <- get_platforms()
+regions   <- get_regions()
+
+use_data(platforms, regions, overwrite = TRUE)